+++ lib/components/fields/selection-field/index.js
@@ -22,6 +22,10 @@ var _fuzzy = require('fuzzy');
 
 var _fuzzy2 = _interopRequireDefault(_fuzzy);
 
+var _extractComponent = require('../../../utils/extract-component');
+
+var _extractComponent2 = _interopRequireDefault(_extractComponent);
+
 var _errors = require('../common/errors');
 
 var _errors2 = _interopRequireDefault(_errors);
@@ -48,29 +52,20 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
 // Import styles
 
 
-function extractComponent(components, name) {
-  var component = false;
-  components.forEach(function (c) {
-    if (c.name === name) {
-      component = c.component;
-    }
-  });
-  return component;
-}
-
-// DefaultSelected
-var SelectedDefault = function SelectedDefault(_ref) {
+/**
+ * Default component for representing a "selected/selection" item
+ * @param  {Object} props Taking the shape of:
+ *
+ * {
+ *   option: { label: 'foo'}
+ *  }
+ *
+ * I.e., expecting the option to have a `label` key with a string value.
+ *
+ * @return {ReactElement}
+ */
+var SelectDefault = function SelectDefault(_ref) {
   var option = _ref.option;
-  return _react2.default.createElement(
-    'div',
-    null,
-    option.label
-  );
-};
-
-// DefaultSelector
-var SelectionDefault = function SelectionDefault(_ref2) {
-  var option = _ref2.option;
   return _react2.default.createElement(
     'div',
     null,
@@ -80,6 +75,9 @@ var SelectionDefault = function SelectionDefault(_ref2) {
 
 /**
  * Selection field
+ *
+ * Handles a singular select of a set of pre-supplied options.
+ *
  */
 var SelectionField = _react2.default.createClass({
   displayName: 'SelectionField',
@@ -105,24 +103,26 @@ var SelectionField = _react2.default.createClass({
     value: _react2.default.PropTypes.oneOfType([_react2.default.PropTypes.string, _react2.default.PropTypes.number])
   },
 
+  /**
+   * Set the default props
+   * @return {Object} Props with defaults
+   */
   getDefaultProps: function getDefaultProps() {
     return {
       placeholder: 'Make a selection',
       select_button_text: 'Select'
     };
   },
+
+
+  /**
+   * Default state, blank search
+   * @return {Object}
+   */
   getInitialState: function getInitialState() {
     return {
-      search: null,
-      options: (0, _immutable.List)(this.props.attributes.options)
+      search: null
     };
-  },
-  componentWillReceiveProps: function componentWillReceiveProps(nextProps) {
-    if (nextProps.attributes.options) {
-      this.setState({
-        options: (0, _immutable.List)(nextProps.attributes.options)
-      });
-    }
   },
 
 
@@ -136,32 +136,82 @@ var SelectionField = _react2.default.createClass({
       return value;
     });
   },
+
+
+  /**
+   * On choose click, open selector
+   * @return {Null}
+   */
   onChooseClick: function onChooseClick(e) {
     e.preventDefault();
     this.openSelector();
   },
+
+
+  /**
+   * When selected item is removed
+   * @return {Null}
+   */
   onRemoveClick: function onRemoveClick(e) {
     e.preventDefault();
     this.onChange(null);
   },
+
+
+  /**
+   * When a selection is made, trigger change and close the selector
+   * @return {Null}
+   */
   onSelection: function onSelection(id) {
     this.closeSelector();
     this.onChange(id);
   },
+
+
+  /**
+   * Open the selector popout
+   * @return {Null}
+   */
   openSelector: function openSelector() {
     this.refs.selector.openPopout();
   },
+
+
+  /**
+   * Close the selector popout
+   * @return {Null}
+   */
   closeSelector: function closeSelector() {
     this.refs.selector.closePopout();
   },
+
+
+  /**
+   * On popout close, reset the search
+   * @return {Null}
+   */
   onPopoutClose: function onPopoutClose() {
     this.setState({
       search: null
     });
   },
+
+
+  /**
+   * On popout open, focus the search input
+   * @return {Null}
+   */
   onPopoutOpen: function onPopoutOpen() {
     this.refs.search.focus();
   },
+
+
+  /**
+   * Fired when search input is `change`d.
+   * Set this.state.search to the value of the input
+   * @param  {Event} e Keyboard event
+   * @return {Null}
+   */
   onSearchChange: function onSearchChange(e) {
     var search = e.target.value;
     this.setState({
@@ -179,9 +229,8 @@ var SelectionField = _react2.default.createClass({
     var label = _props.label;
     var name = _props.name;
     var value = _props.value;
-    var _state = this.state;
-    var options = _state.options;
-    var search = _state.search;
+    var search = this.state.search;
+    var options = attributes.options;
     var placeholder = attributes.placeholder;
     var select_button_text = attributes.select_button_text;
     var selected_component = attributes.selected_component;
@@ -193,16 +242,16 @@ var SelectionField = _react2.default.createClass({
     var fieldClassNames = (0, _classnames2.default)(_selectionField2.default.base, _defineProperty({}, '' + _selectionField2.default.baseInline, attributes.inline));
 
     // Determine the selection/selected display components
-    var Selected = SelectedDefault;
-    var Selection = SelectionDefault;
+    var Selected = SelectDefault;
+    var Selection = SelectDefault;
 
     // Extract them from the passed `config.components` if it exists
     if (config.components) {
       if (selected_component) {
-        Selected = extractComponent(config.components, selected_component) || Selected;
+        Selected = (0, _extractComponent2.default)(config.components, selected_component) || Selected;
       }
       if (selection_component) {
-        Selection = extractComponent(config.components, selection_component) || Selection;
+        Selection = (0, _extractComponent2.default)(config.components, selection_component) || Selection;
       }
     }
 
@@ -237,6 +286,7 @@ var SelectionField = _react2.default.createClass({
         _react2.default.createElement(Selection, { option: option })
       );
     });
+
     return _react2.default.createElement(
       'div',
       { className: fieldClassNames },
@@ -300,7 +350,7 @@ var SelectionField = _react2.default.createClass({
               _react2.default.createElement(
                 'div',
                 { className: _selectionField2.default.selectionsList },
-                selections.count() > 0 ? selections : _react2.default.createElement(
+                selections.length > 0 ? selections : _react2.default.createElement(
                   'p',
                   { className: _selectionField2.default.noResults },
                   'No matching results'
